requirements.txt
joblib==1.3.2
python-dotenv==1.0.1
minio==7.2.3
s3fs==2023.12.2
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
pendulum==3.0.0
plotly==5.18.0
python-dotenv==1.0.1
scikit_learn==1.4.0
textblob==0.18.0.post0
tweepy==4.14.0
unicorn_binance_rest_api==2.2.0
vaderSentiment==3.3.2
mlflow==2.11.1
tensorflow==2.15.0.post1
